+++ Revit/Revit2023/Speckle.Revit2023.Interfaces/Speckle.Revit2023.Interfaces.csproj
@@ -3,6 +3,7 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
   <Import Project="..\..\Speckle.Revit.Interfaces\Speckle.Revit.Interfaces.projitems" Label="Shared" />
+  <Import Project="..\..\..\Speckle.Shared\Speckle.Shared.projitems" Label="Shared" />
   <ItemGroup>
     <PackageReference Include="MinVer" PrivateAssets="All"/>
     <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
