+++ Revit/Speckle.Revit.Api/RevitUnitUtils.cs
@@ -2,6 +2,7 @@
 using Mapster.Utils;
 using Speckle.ProxyGenerator;
 using Speckle.Revit.Interfaces;
+using Speckle.Shared;
 
 namespace Speckle.Revit.Api;
