+++ Rhino/7/Speckle.Rhino7.Interfaces/Speckle.Rhino7.Interfaces.csproj
@@ -3,6 +3,7 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
   <Import Project="..\..\Speckle.Rhino.Interfaces\Speckle.Rhino.Interfaces.projitems" Label="Shared" />
+  <Import Project="..\..\..\Speckle.Shared\Speckle.Shared.projitems" Label="Shared" />
   <ItemGroup>
     <PackageReference Include="MinVer" PrivateAssets="All"/>
     <PackageReference Include="Microsoft.SourceLink.GitHub" PrivateAssets="All" />
