+++ Rhino/Speckle.Rhino.Interfaces/RhinoUnitSystem.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections.Concurrent;
-
-namespace Speckle.Rhino7.Interfaces;
+﻿namespace Speckle.Rhino7.Interfaces;
 
 public enum RhinoUnitSystem : byte
 {
@@ -121,192 +119,4 @@ public enum RhinoUnitSystem : byte
   ///  This is a differnt condition from ON::LengthUnitSystem::None.
   /// </summary>
   Unset = 255, // 0xFF
-}
-
-public enum RhinoLoopFormula
-{
-  /// <summary>The original Loop formula.</summary>
-  Loop,
-
-  /// <summary>Warren's improved formula.</summary>
-  Warren,
-
-  /// <summary>
-  /// The formula as explained by Weimer, improving over Warren's.
-  /// </summary>
-  WarrenWeimer,
-}
-
-public enum RhinoBrepSolidOrientation
-{
-  /// <summary>Brep is a Solid with inward facing normals.</summary>
-  Inward = -1, // 0xFFFFFFFF
-
-  /// <summary>Brep is not a Solid.</summary>
-  None = 0,
-
-  /// <summary>Brep is a Solid with outward facing normals.</summary>
-  Outward = 1,
-
-  /// <summary>
-  /// Breps is a Solid but no orientation could be computed.
-  /// </summary>
-  Unknown = 2,
-}
-
-public enum RhinoIsoStatus
-{
-  /// <summary>curve is not an isoparametric curve.</summary>
-  None,
-
-  /// <summary>
-  /// curve is a "x" = constant (vertical) isoparametric curve in the interior of the surface's domain.
-  /// </summary>
-  X,
-
-  /// <summary>
-  /// curve is a "y" = constant (horizontal) isoparametric curve in the interior of the surface's domain.
-  /// </summary>
-  Y,
-
-  /// <summary>
-  /// curve is a "x" = constant isoparametric curve along the west side of the surface's domain.
-  /// </summary>
-  West,
-
-  /// <summary>
-  /// curve is a "y" = constant isoparametric curve along the south side of the surface's domain.
-  /// </summary>
-  South,
-
-  /// <summary>
-  /// curve is a "x" = constant isoparametric curve along the east side of the surface's domain.
-  /// </summary>
-  East,
-
-  /// <summary>
-  /// curve is a "y" = constant isoparametric curve along the north side of the surface's domain.
-  /// </summary>
-  North,
-}
-
-public enum RhinoBrepTrimType
-{
-  /// <summary>Unknown type</summary>
-  Unknown,
-
-  /// <summary>
-  /// Trim is connected to an edge, is part of an outer, inner or
-  /// slit loop, and is the only trim connected to the edge.
-  /// </summary>
-  Boundary,
-
-  /// <summary>
-  /// Trim is connected to an edge, is part of an outer, inner or slit loop,
-  /// no other trim from the same loop is connected to the edge, and at least
-  /// one trim from a different loop is connected to the edge.
-  /// </summary>
-  Mated,
-
-  /// <summary>
-  /// trim is connected to an edge, is part of an outer, inner or slit loop,
-  /// and one other trim from the same loop is connected to the edge.
-  /// (There can be other mated trims that are also connected to the edge.
-  /// For example, the non-manifold edge that results when a surface edge lies
-  /// in the middle of another surface.)  Non-manifold "cuts" have seam trims too.
-  /// </summary>
-  Seam,
-
-  /// <summary>
-  /// Trim is part of an outer loop, the trim's 2d curve runs along the singular
-  /// side of a surface, and the trim is NOT connected to an edge. (There is
-  /// no 3d edge because the surface side is singular.)
-  /// </summary>
-  Singular,
-
-  /// <summary>
-  /// Trim is connected to an edge, is the only trim in a curve-on-surface loop, and
-  /// is the only trim connected to the edge.
-  /// </summary>
-  CurveOnSurface,
-
-  /// <summary>
-  /// Trim is a point on a surface, trim.m_pbox is records surface parameters,
-  /// and is the only trim in a point-on-surface loop.  This trim is not connected to
-  /// an edge and has no 2d curve.
-  /// </summary>
-  PointOnSurface,
-
-  /// <summary></summary>
-  Slit,
-}
-
-public enum RhinoBrepLoopType
-{
-  /// <summary>Unknown loop type.</summary>
-  Unknown,
-
-  /// <summary>
-  /// 2d loop curves form a simple closed curve with a counterclockwise orientation.
-  /// </summary>
-  Outer,
-
-  /// <summary>
-  /// 2d loop curves form a simple closed curve with a clockwise orientation.
-  /// </summary>
-  Inner,
-
-  /// <summary>
-  /// Always closed - used internally during splitting operations.
-  /// </summary>
-  Slit,
-
-  /// <summary>
-  /// "loop" is a curve-on-surface made from a single (open or closed) trim that
-  /// has type TrimType.CurveOnSurface.
-  /// </summary>
-  CurveOnSurface,
-
-  /// <summary>
-  /// "loop" is a PointOnSurface made from a single trim that has
-  /// type TrimType.PointOnSurface.
-  /// </summary>
-  PointOnSurface,
-}
-
-public static class EnumUtility<TSource, TDestination>
-  where TSource : struct, Enum
-  where TDestination : struct, Enum
-{
-  private static readonly ConcurrentDictionary<TSource, TDestination> _destinations = new();
-
-  static EnumUtility()
-  {
-    var sources = ((TSource[])Enum.GetValues(typeof(TSource))).Select(x => (x.ToString(), x));
-    var destinations = Enum.GetNames(typeof(TDestination))
-      .Select(x => (x, (TDestination)Enum.Parse(typeof(TDestination), x.ToString())))
-      .ToList();
-    foreach (var (name, val) in sources)
-    {
-      var d = destinations.Where(x => x.Item1 == name).ToList();
-      if (d.Any())
-      {
-        foreach (var v in d)
-        {
-          destinations.Remove(v);
-        }
-        _destinations.TryAdd(val, d.Single().Item2);
-      }
-    }
-  }
-
-  public static TDestination Convert(TSource source)
-  {
-    if (_destinations.TryGetValue(source, out var destination))
-    {
-      return destination;
-    }
-
-    throw new InvalidOperationException($"{source} does not exist in destination enum: {typeof(TDestination)}");
-  }
 }
