+++ Speckle.Host.Apis.sln
@@ -66,6 +66,10 @@ Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Speckle.Rhino.Api", "Rhino\
 EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Speckle.Rhino.Interfaces", "Rhino\Speckle.Rhino.Interfaces\Speckle.Rhino.Interfaces.shproj", "{E1C43415-3200-45F4-8BF9-A4DD7D7F2ED6}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{3B0D8F6E-C701-446E-BAE3-B6FDF1812901}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Speckle.Shared", "Speckle.Shared\Speckle.Shared.shproj", "{E1C43415-3200-45F4-8BF9-B4DD7D7F2ED6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -118,6 +122,7 @@ Global
 		{BC12C4CA-7065-4B76-9787-E8DA4745AD4E} = {72DEFF6C-31DE-4050-8692-AEE8224E9925}
 		{EB2B9350-BE2B-4197-ACA8-3963EAE3B6C9} = {BC12C4CA-7065-4B76-9787-E8DA4745AD4E}
 		{5F678DE0-AAE7-4FA9-9A7F-C3C0B0E3E3D7} = {BC12C4CA-7065-4B76-9787-E8DA4745AD4E}
-		{E1C43415-3200-45F4-8BF9-A4DD7D7F2ED6} = {8F540B7F-7548-46E9-BDEC-4DAB99367908}
+		{E1C43415-3200-45F4-8BF9-B4DD7D7F2ED6} = {3B0D8F6E-C701-446E-BAE3-B6FDF1812901}
+		{E1C43415-3200-45F4-8BF9-A4DD7D7F2ED6} = {72DEFF6C-31DE-4050-8692-AEE8224E9925}
 	EndGlobalSection
 EndGlobal
